requirements.txt
matplotlib
pandas
python-dotenv
seaborn
sklearn
streamlit